fix: findFastestAnsweredChallenge by Lucky
