+++ src/main/java/ongjong/namanmoo/service/ChallengeServiceImpl.java
@@ -350,116 +350,4 @@ public class ChallengeServiceImpl implements ChallengeService {
         return null;
     }
 
-    // 모든 가족 구성원이 가장 빠르게 답한 챌린지를 반환
-    @Override
-    public Challenge findFastestAnsweredChallenge(Lucky lucky) throws Exception {
-        Family family = lucky.getFamily();
-        List<Challenge> challenges = findRunningChallenges(); // 해당 Lucky의 모든 챌린지 가져오기
-        Challenge fastestChallenge = null;
-        long shortestTime = Long.MAX_VALUE;
-
-        for (Challenge challenge : challenges) {
-            List<Challenge> relatedChallenges = challengeRepository.findByChallengeNum(challenge.getChallengeNum());
-            int totalMembers = family.getMembers().size();
-
-            // 모든 관련 챌린지에 대해 모든 가족 구성원이 답변을 완료했는지 확인
-            boolean allRelatedChallengesAnswered = true;
-            for (Challenge relatedChallenge : relatedChallenges) {
-                List<Answer> relatedAnswers = answerRepository.findByChallengeAndMemberFamily(relatedChallenge, family);
-                if (relatedAnswers.size() != totalMembers) {
-                    allRelatedChallengesAnswered = false;
-                    break;
-                }
-            }
-
-            if (!allRelatedChallengesAnswered) {
-                continue;
-            }
-
-            // 각 챌린지에 대해 답변 확인 및 가장 늦은 응답 시간 계산
-            for (Challenge relatedChallenge : relatedChallenges) {
-                List<Answer> answers = answerRepository.findByChallengeAndMemberFamily(relatedChallenge, family);
-                log.info("Challenge ID: " + relatedChallenge.getChallengeId() + ", Answer count: " + answers.size());
-
-                // 해당 챌린지에 대해 모든 가족 구성원이 답변 내용을 입력했는지 확인
-                boolean allMembersAnswered = true;
-                for (Answer answer : answers) {
-                    if (answer.getAnswerContent() == null || answer.getAnswerContent().isEmpty() || answer.getModifiedDate() == null || answer.getModifiedDate().isEmpty()) {
-                        allMembersAnswered = false;
-                        break;
-                    }
-                }
-
-                if (allMembersAnswered) {
-                    long timeToAnswer = calculateLatestResponseTime(lucky, relatedChallenge); // 챌린지별 가장 늦은 응답 시간 계산
-                    log.info("Challenge ID: " + relatedChallenge.getChallengeId() + ", Time to answer: " + timeToAnswer);
-                    if (timeToAnswer < shortestTime) {
-                        shortestTime = timeToAnswer;
-                        fastestChallenge = relatedChallenge;
-                    }
-                }
-            }
-        }
-        log.info("Fastest Challenge ID: " + (fastestChallenge != null ? fastestChallenge.getChallengeId() : "None"));
-        return fastestChallenge;
-    }
-
-    // 챌린지에 달린 답변 중, 가장 늦게 달린 답변 시간을 계산
-    @Override
-    public long calculateLatestResponseTime(Lucky lucky, Challenge challenge) throws Exception {
-        long latestTime = Long.MIN_VALUE; // 해당 챌린지의 가장 늦은 응답시간을 저장할 변수
-        List<Answer> answers = answerRepository.findByChallenge(challenge);
-
-        SimpleDateFormat format4 = new SimpleDateFormat(DateUtil.FORMAT_4);
-        SimpleDateFormat format9 = new SimpleDateFormat(DateUtil.FORMAT_9);
-
-        for (Member member : lucky.getFamily().getMembers()) {
-            long latestResponseTimeForMember = Long.MIN_VALUE; // 가족 구성원의 가장 늦은 응답시간을 저장할 변수
-
-            for (Answer answer : answers) {
-                if (answer.getMember().equals(member)) {
-                    try {
-                        String createDate = answer.getCreateDate();
-                        String modifiedDate = answer.getModifiedDate();
-
-                        if (createDate == null || modifiedDate == null) {
-                            log.warn("답변 ID: {}에서 Null 타임스탬프를 찾았습니다.", answer.getAnswerId());
-                            continue;
-                        }
-
-                        Date createTime = format4.parse(createDate);
-                        Date modifiedTime = format9.parse(modifiedDate);
-                        log.info("답변 수정 시간: {}, 답변 생성 시간: {}", modifiedTime.getTime(), createTime.getTime());
-
-                        long responseTime = Math.abs(modifiedTime.getTime() - createTime.getTime()); // 원래라면 수정 날짜가 답변이 열리는 날짜보다 나중이기 때문에 음수가 나올일이 없지만 올바른 시연을 위해 절대값 처리
-                        log.info("회원 ID: {}, 응답 시간: {}", member.getMemberId(), responseTime);
-
-                        if (responseTime > latestResponseTimeForMember) {
-                            latestResponseTimeForMember = responseTime;
-                        }
-                    } catch (ParseException e) {
-                        log.error("답변 ID: {}의 날짜 파싱 에러", answer.getAnswerId(), e);
-                    }
-                }
-            }
-
-            if (latestResponseTimeForMember > latestTime) {
-                latestTime = latestResponseTimeForMember;
-            }
-        }
-
-        if (latestTime == Long.MIN_VALUE) {
-            return Long.MAX_VALUE;
-        }
-        long totalSeconds = latestTime / 1000;
-        long totalMinutes = totalSeconds / 60;
-        long totalHours = totalMinutes / 60;
-        long days = totalHours / 24;
-        long seconds = totalSeconds % 60;
-        long minutes = totalMinutes % 60;
-        long hours = totalHours % 24;
-        log.info("가장 늦은 응답 시간: " + days + "일 " + hours + "시간 " + minutes + "분 " + seconds + "초");
-        return latestTime;
-    }
-
 }
